melhoria da paginacao e objeto de response do find all

+++ src/main/java/br/com/paulomoreira/starB2W/resource/PlanetService.java
@@ -4,15 +4,19 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
 import br.com.paulomoreira.starB2W.dto.PlanetRequest;
+import br.com.paulomoreira.starB2W.exception.AttributeException;
 import br.com.paulomoreira.starB2W.gateway.SwapiGateway;
 import br.com.paulomoreira.starB2W.gateway.dto.SwapiResponse;
 import br.com.paulomoreira.starB2W.model.Planet;
+import br.com.paulomoreira.starB2W.resource.repository.PlanetRepository;
 import br.com.paulomoreira.starB2W.util.Constants;
 import br.com.paulomoreira.starB2W.util.Converter;
 import br.com.paulomoreira.starB2W.util.Validation;
@@ -35,20 +39,34 @@ public class PlanetService {
 	@Autowired
 	Validation validation;
 
-	public List<Planet> findAllPlanets(Integer page, Integer size) {
+	public Page<Planet> findAllPlanets(Integer page) {
 
 		log.info(("Searching all the planets."));
-		Pageable pageable = this.creatingPagination(page, size);
-		List<Planet> planets = planetRepository.findAll(pageable).getContent();
 
-		if (planets.isEmpty()) {
-			planets = planetRepository.findAll();
-			if (planets.isEmpty()) {
-				return planets;
-			}
+		if (page == null) {
+
+			List<Planet> planets = planetRepository.findAll();
+			Page<Planet> planetsPage = new PageImpl<Planet>(planets);
+
+			return planetsPage;
+
 		}
 
-		return planets;
+		page = page - 1;
+
+		Pageable pageable = this.creatingPagination(page, 10);
+		Page<Planet> planetsPage = planetRepository.findAll(pageable);
+
+		if (planetsPage.getContent().isEmpty()) {
+			List<Planet> planets = planetRepository.findAll();
+
+			planetsPage = new PageImpl<Planet>(planets);
+
+			if (planets.isEmpty()) {
+				return planetsPage;
+			}
+		}
+		return planetsPage;
 
 	}
 
@@ -85,17 +103,26 @@ public class PlanetService {
 
 	}
 
-	public Optional<Planet> createPlanet(PlanetRequest planetRequest) {
+	public Optional<Planet> createPlanet(PlanetRequest planetRequest) throws AttributeException {
 
 		String planetName = StringUtils.capitalize(planetRequest.getName());
 
-		Boolean planetExistsInDatebase = validation.checkIfPlanetExistInDatabase(planetName);
+		Boolean planetExistsInDatebase = validation.checkIfPlanetExistInDatabaseByName(planetName);
 
 		if (planetExistsInDatebase == Constants.FALSE) {
 
 			String movieAppearances = this.checkForMovieAppearances(planetName);
-
 			Optional<Planet> planet = converter.requestToPlanet(planetRequest, movieAppearances);
+
+			Boolean climateIsValid = validation.climate(planetRequest.getClimates());
+			Boolean terrainIsValid = validation.terrain(planetRequest.getTerrains());
+
+			if (climateIsValid == false) {
+				throw new AttributeException(Constants.CLIMATE_INVALID);
+			}
+			if (terrainIsValid == false) {
+				throw new AttributeException(Constants.TERRAIN_INVALID);
+			}
 
 			log.info("Creating planet {}.", planet.get().getName());
 			planet = Optional.of(planetRepository.save(planet.get()));
@@ -110,7 +137,7 @@ public class PlanetService {
 
 	public Optional<Planet> createPlanetByName(String planetName) {
 
-		Boolean planetExists = validation.checkIfPlanetExistInDatabase(planetName);
+		Boolean planetExists = validation.checkIfPlanetExistInDatabaseByName(planetName);
 
 		if (planetExists == false) {
 
@@ -121,12 +148,15 @@ public class PlanetService {
 					return null;
 				}
 
+				List<String> climates = converter.climateToList(response.getResults().get(0).getClimate());
+				List<String> terrains = converter.terrainsToList(response.getResults().get(0).getTerrain());
 				Optional<Planet> planet = Optional.of(new Planet());
 
 				planet.get().setName(response.getResults().get(0).getName());
-				planet.get().setClimate(response.getResults().get(0).getClimate());
-				planet.get().setTerrain(response.getResults().get(0).getTerrain());
-
+				planet.get().setClimates(climates);
+				planet.get().setTerrains(terrains);
+				String movieAppearances = String.valueOf(response.getResults().get(0).getFilms().size());
+				planet.get().setMovieAppearances(movieAppearances);
 				planet = Optional.of(planetRepository.save(planet.get()));
 
 				return planet;
@@ -137,8 +167,8 @@ public class PlanetService {
 			}
 
 		}
-		 Optional<Planet> emptyPlanet = Optional.empty();
-		 return emptyPlanet;
+		Optional<Planet> emptyPlanet = Optional.empty();
+		return emptyPlanet;
 
 	}
 
@@ -160,13 +190,13 @@ public class PlanetService {
 
 	public Boolean deletePlanById(String id) {
 
-		log.info("Deleting planet with Id{}.", id);
-		Long isDeleted = planetRepository.deleteById(id);
+		Boolean planetExists = validation.checkIfPlanetExistInDatabaseById(id);
 
-		if (isDeleted == Constants.ONE) {
-			log.info("Planet with Id {} deleted.", id);
-			return Constants.TRUE;
+		if (planetExists == true) {
+			log.info("Deleting planet with Id{}.", id);
+			planetRepository.deleteById(id);
 		}
+
 		log.info("The planet with Id {} was not found.", id);
 		return Constants.FALSE;
